requirements.txt
flask
opencv-python
deepface
gtts
numpy
pymysql
gradio
pillow